pyproject.toml
[tool.poetry]
name = "mars-rover"
version = "0.1.0"
description = "Mars Rover exercise"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.14.0"
Flask = "^2.1.1"
requests = "^2.28.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
